requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2

# ✅ 추가된 패키지
requests>=2.31.0
scikit-learn>=1.5.0
openpyxl>=3.1.2
gunicorn>=21.2.0
konlpy>=0.6.0
Pillow>=10.3.0